Set up for compiling with Visual Studio & fix associated warnings

+++ src/lobject.h
@@ -782,7 +782,7 @@ typedef struct Table {
 	(check_exp((size&(size-1))==0, (cast_int((s) & ((size)-1)))))
 
 
-#define twoto(x)	(1<<(x))
+#define twoto(x)	(int)(1<<(x))
 #define sizenode(t)	(twoto((t)->lsizenode))
